Don't qualify import for Control.Applicative.many
We used to use A.many but attoparsec 0.10 and up no longer
exports it (basically a re-export with some performance tweaks).

Commit e0269eb8d74f1688acade5f7788092e08faa6041 switches this
to Control.Applicative.many.

Commit 68fe5ba998adcdef2e3cee6712a54ec586f3bde8 unthinkingly
uses an unqualified import. This turns out to actually be not
a bad way to do it. Since we import attoparsec is imported
qualified anyway there's no risk of ambiguity if we happen to
be built against an older attoparsec.