Support escapes when parsing symbol names.
The approach here is to focus on accepting symbol names rather than
trying to interpret the escapes, etc.  So what the user gets back will
have all the backslashes, | symbols etc in it and it's up to them to
figure out the case sensitivity issues themselves.  The same goes for
package name prefixes.